+++ src/main/java/moze_intel/projecte/gameObjs/container/TransmutationContainer.java
@@ -92,7 +92,7 @@ public class TransmutationContainer extends Container
 		}
 		else if (slotIndex >= 10 && slotIndex <= 25) // Output Slots
 		{	
-			int emc = EMCHelper.getEmcValue(newStack);
+			long emc = EMCHelper.getEmcValue(newStack);
 			
 			int stackSize = 0;
 			
@@ -107,7 +107,7 @@ public class TransmutationContainer extends Container
 		}
 		else if (slotIndex >= 26) //Unlearn Slot and Player Inventory
 		{
-			int emc = EMCHelper.getEmcValue(stack);
+			long emc = EMCHelper.getEmcValue(stack);
 			
 			if (emc == 0 && stack.getItem() != ObjHandler.tome)
 			{
